adding and testing some Lua sripting for NPC and other events and will be modifying the engine to suit the needs of the Android doom-ish engine

+++ Source/Core/Config/ProgramConfiguration.cs
@@ -336,6 +336,11 @@ namespace CodeImp.DoomBuilder.Config
             {
                 // Copy new configuration
                 General.WriteLogLine("Local user program configuration is missing!");
+                if (!File.Exists(defaultfilepathname))
+                {
+                    General.WriteLogLine("Default program configuration is also missing: " + defaultfilepathname);
+                    return false;
+                }
                 File.Copy(defaultfilepathname, cfgfilepathname);
                 General.WriteLogLine("New program configuration copied for local user");
             }
